Debugging and testing - 7, final fixes

+++ Backend/Migrations/SQL Tables/007_create_waste_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS waste_collection(
-    id BIGINT NOT NULL PRIMARY KEY,
+    id SERIAL NOT NULL PRIMARY KEY,
     userid UUID REFERENCES users(id) NOT NULL,
     location TEXT NOT NULL,
     weight DECIMAL NOT NULL,
